Optimize `ProtocolFeeLibrary` (#283)
* Refactor import statements and add new gas test

Refactored import statements in ProtocolFeeLibrary.t.sol and added a new test in PoolManager.t.sol to check the gas usage when setting the protocol fee. The changes in import statements are done to follow the file path consistency and the new gas test will help in stress testing to evaluate the gas cost for a particular operation.

* Optimize `ProtocolFeeLibrary.validate`

Refactored and optimized the `validate` function in the `ProtocolFeeLibrary.sol` file. The logic for validating the fee has been simplified and compressed using assembly language. This results in a reduction in bytecode sizes in several snapshots including.poolManager, initialize, and set protocol fee.

* Add gas test for `calculateSwapFee`

This commit introduces a gas test for the `calculateSwapFee` function inside the `ProtocolFeeLibraryTest` contract. With the addition of `GasSnapshot` import, it now tracks the gas consumption of this specific function, helping us to optimize efficiency and performance.

* Refactor `calculateSwapFee` function to use assembly

The `calculateSwapFee` function in the `ProtocolFeeLibrary.sol` file has been rewritten to use assembly for performance optimization. This improves efficiency by reducing gas usage, reflected in the bytecode size reduction, which is captured by the snapshot changes. Operating at a lower level with assembly allows us to manage resources more directly.

* Change visibility of protocol fee thresholds

* Optimize `isValidProtocolFee` further

* Apply review comments