Refactor and optimize `SwapMath` (#280)
* Refactor and optimize `SwapMath`

* Update `SwapMath` calculations and gas snapshots, change `test_entireInputAmountTakenAsFee`

* No assembly in `computeSwapStep`

Replace inline assembly with native Solidity for `amountRemainingAbs` calculation in `computeSwapStep` function. The previous implementation was not intuitive and easy to read. This change simplifies the code and improves readability while maintaining the exact functionality.

* Avoid casting `feePips` multiple times

In `SwapMath.sol`, we reduced redundancy by storing the value of `feePips` into `_feePips` and using it throughout the function instead of casting `feePips` multiple times. This refactoring improved efficiency and reduced bytecode size, leading to reduced gas cost for function calls.

* Fix natspec

* Rename `sqrtRatio` to `sqrtPrice` in SwapMath and Pool libraries

* Refactor variable declaration in `SwapMath`

The declaration and assignment of `amountRemainingAbs` has been moved inside the relevant if-else blocks in the SwapMath.sol file. This change makes the code more concise and clear by limiting the scope and use of `amountRemainingAbs` to where it is needed.

* Update test function to pure in SwapMath.t.sol

* Optimize `getSqrtPriceTarget`

* Switch if else blocks for ease of review

* Refactor and reduce diffs

* Recompute `amountIn` when it's capped by `amountRemaining` instead of the target price for `exactIn` to achieve equivalence with V3

* Remove `amountRemainingAbs`

* Remove .gas-snapshot

* Address review comment

* fmt